growpart: avoid race in partition table reload
udevd watches files in /dev via inotify, and when a block disk device
is closed after writing, it tells the kernel to reload the partition
table.

This was racing with the reload code in growpart; if we were unlucky
in our timing, the kernel would be busy with udevd's reload and our
call would fail.

To fix this, we stop reloading the partition table in growpart, and
use inotify to determine when udevd has deleted and recreated the
partition links.

Signed-off-by: Ben Cressey <bcressey@amazon.com>