pubsys: use base region for contacting STS
The region used for the base credentials provider should be the one in which
you want to talk to STS to get temporary credentials, not the region in which
you want to talk to a service endpoint like EC2.  This is needed because you
may be assuming a role in an opt-in region from an account that has not
opted-in to that region, and you need to get session credentials from an STS
endpoint in a region to which you have access in the base account.