build: align opt-level for crate build deps
Our OS packages are built as a byproduct of build script execution,
while the final Rust artifacts are largely ignored.

Since Rust 1.47.0, cargo builds host dependencies like build scripts
with opt-level "0" by default, which did not match the "z" level we
picked to minimize the size of the output artifacts.

This caused package build dependencies to be built more than once,
with concurrent Docker builds for the same package. This is wasteful
and creates races that our build tool does not handle gracefully.

By aligning the opt-level we restore the previous behavior.

Signed-off-by: Ben Cressey <bcressey@amazon.com>