create links for ephemeral storage devices
Some EC2 instance types come with disks that are physically connected
to the host. Kubernetes can use these as local persistent volumes in
conjunction with an external static provisioner.

To make it easier to configure the static provisioner, we now create
symlinks to known ephemeral disk devices in /dev/disk/ephemeral.

Incorrectly identifying a device as an ephemeral when it is not could
cause irrecoverable data loss. Hence we take an "allowlist" approach
where only specific device types are included, and check for known
partition types as an additional safeguard.