chore: Refactor `amiize` script.
When developing https://github.com/bottlerocket-os/bottlerocket/pull/1160 running the `amiize` script failed on my machine repeatedly (macOS 10.15.7) using `aws-vault` in the `us-west-2` region. The script itself failed in then tried to continue due to lack of `set -o pipefile` and so a set of more strict shell options has been added as a prologue. Because of this, `check_return` isn't needed per se since the script just fails if the command fails. The file size calculator also fails to work on macOS because the `du` command is missing some (presumably GNU) flags; so instead we use `awk`, `ls` and `bc` to pull out the size in GiB.

These refactorings are not at all required and were just useful for my development and testing and I am in no way suggesting or asking you incorporate them, but if there's something useful you find in here feel free to take it!