apiserver: do not block on sd_notify status
In v246, systemd added an sd_notify_barrier function to avoid a race
when the notification is sent by a different process than the one it
spawned. The barrier works by sharing a file descriptor between the
process and systemd.

That doesn't apply to our use case, and the shared file descriptor
complicates the SELinux policy, since systemd would need to write to
a file owned by apiserver.

Work around this by passing the new `--no-block` argument to the
`systemd-notify` command.

Signed-off-by: Ben Cressey <bcressey@amazon.com>