build: add target for kmod kit
To support compiling out-of-tree modules ahead of time, rather than
on a running Bottlerocket host, we need to provide two things: the
kernel development sources, such as headers and Makefiles; and the
toolchain we use to build our kernel.

Our toolchain is built separately as part of our cross-compiling SDK,
and it's possible, if unlikely, that we would ship two releases with
the same kernel version built with a different GCC.

It's also possible that variants will use different kernels, so we
cannot have just one development kit per release. This is not yet
supported, but we need the ecosystem to anticipate the requirement
for a per-variant, per-architecture kit.

The build target combines the archives from the toolchain matching
the SDK we used to build the kernel, and kernel development sources
from the most recent build. This produces a single artifact that
can be uploaded for later retrieval by a consumer that knows the
variant, architecture, and version that they are targeting.

Signed-off-by: Ben Cressey <bcressey@amazon.com>