thar-be-updates: manage tokio runtime manually so we don't fork threads
The `refresh` command needs to talk to the API to get the current version-lock
setting.  390aafe3ebf made apiclient::raw_request an async call and changed
thar-be-settings to use tokio::main, but didn't consider that the async call is
happening after a fork in thar-be-updates.  In general it's not safe to use the
threaded tokio runtime over a fork, and it caused thar-be-updates to hang,
holding its lock file forever.

This undoes the changes from 390aafe3ebf for thar-be-updates so it doesn't have
a global runtime, and instead creates a runtime just for the single async call
we need to make, after the fork.