release: enable writes to /usr/src/kernels
With the changes to our packaging of kernel development sources, any
out-of-tree module builds will need to run `make prepare` in order to
compile dependencies like `objtool`. These binaries need to land in
the same directory tree as the other development files we ship.

Using an overlayfs mount allows writes to the otherwise read-only
content from the squashfs. We purge the upper directory on reboot so
changes do not persist across system upgrades.

Signed-off-by: Ben Cressey <bcressey@amazon.com>