selinux-policy: adjust initial SIDs for network entities
We don't use any of the SELinux controls for network access, such as
packet or peer labeling. This sets permissive initial SIDs for these
entities to avoid unexpected restrictions.

Clean up the treatment of other initial SIDs at the same time. Many
of the unused ones have network-related names, and the fact that they
are treated differently can lead to confusion.

Signed-off-by: Ben Cressey <bcressey@amazon.com>