selinux-policy: clean up handling of subject types
The intent was for type attributes like "trusted_s" to only contain
subject types, but they were defined by negation and also included
all the object types. By using xor, we end up with the set members
we expect.

Since objects are not typically the actor in an SELinux access check,
this did not result in any additional permissions being granted, with
one exception: the ability to add a file to a filesystem is governed
by the "associate" action on the "filesystem" class, where the actor
is the file rather than the process.

Removing objects from the set of trusted users caused the system to
fail to boot, since files could no longer be created or relabeled.
To fix this, the "associate" action has been removed from the set of
"relabel" actions, and replaced with rules that permit the expected
behavior.

Signed-off-by: Ben Cressey <bcressey@amazon.com>