selinux-policy: revise access to watch actions
These actions were originally restricted because fanotify can be used
to block access to files via permission events, which could disrupt
the operation of the host.

However, fanotify doesn't deliver events for accesses from outside
the mount namespace, so it's unlikely that containers depend on this
functionality.

Meanwhile, the restrictions on file and directory watches interfered
with containers that tried to monitor the journal file in order to
collect logs.

The updated policy allows broad use of inotify and dnotify watches,
while preventing access to fanotify's permission events.

Signed-off-by: Ben Cressey <bcressey@amazon.com>