ecs-agent: update build for 1.48.1
This bump includes changes [^1] that were made to a number of
container images built for the agent. The nested build that was used
to drive the build of `pause` no longer functioned. The build process
now directly builds the executable to sidestep the new build - which
relies directly on Docker.

[^1]: https://github.com/aws/amazon-ecs-agent/commit/7be883bac09b4b60c02ce54d759d1fc14769e289#diff-6ef5a00ce2b6a82260b403d7c14a7ad519480b5f928012a5e3fd355a29310bbe

Signed-off-by: Jacob Vallejo <jakeev@amazon.com>
Tested-By: Shailesh Gothi <gothisg@amazon.com>