Fix build.yml to ignore changes to all .md files
That's a lot of CI to run for a PR to a markdown file. This PR should just say "hey, if it's all md files, just, y'know, don't do that."

I can't think of any cases in which we'd want .md file changes to kick off CI runs, but if you can, feel free to reject this PR.

Also, here's to the GitHub built-in editor for making sure the yaml is right.