kernel: enable early LSM support for lockdown
This fixes an issue where the Lockdown LSM wasn't actually consulted
for security decisions, because it does not appear in the CONFIG_LSM
list.

As an alternative, we could override CONFIG_LSM to add lockdown:

  CONFIG_LSM="lockdown,yama,loadpin,safesetid,integrity,
              selinux,smack,tomoyo,apparmor"

However, then that list would need to be maintained when updating the
kernel, or any new LSMs would be quietly ignored.

Enabling early LSM support resolves the issue, and makes it possible
for custom builds to specify the desired lockdown state on the kernel
command line.

Signed-off-by: Ben Cressey <bcressey@amazon.com>