build: track and clean output files
This adds a layer of indirection between the build artifacts and the
final output directory, so we can keep track of what we've previously
built for a given package or variant.

That allows us to remove the files before the next build, so they do
not interact with other builds in unexpected ways.

Signed-off-by: Ben Cressey <bcressey@amazon.com>