selinux-policy: modify default label for containers
This changes the default label from `container_t`, which is meant as
the unprivileged type, to `control_t`, which has higher privileges,
including the ability to write to the host API socket.

This is better aligned with the SELinux logic in containerd's CRI
plugin and in the docker daemon, where the label from lxc_contexts is
used for unprivileged containers, while privileged containers are not
explicitly labeled and follow the default transition rules.

Signed-off-by: Ben Cressey <bcressey@amazon.com>