docker-engine: drop patch for privileged and SELinux
We added this patch when we were on Docker 18.09.9, which vendored a
copy of go-selinux at commit b6fa367. In that version, passing the
"disable" option would clear both the process and mount labels. A fix
for this was merged into go-selinux at commit b5cb66b, so the mount
label would no longer be cleared. The fix was present in v1.0.0 and
later versions.

When we moved to Docker 19.03.12, the vendored copy of go-selinux had
been updated to v1.3.1, and we no longer needed the patch to ensure
that the root filesystem of a privileged container would be mounted
with the expected label to allow writes.

Signed-off-by: Ben Cressey <bcressey@amazon.com>