metricdog: prepend scheme to https proxy value
reqwest does not support HTTPS_PROXY values that are not in URL format
(i.e. if there is no protocol scheme). On Bottlerocket we expect an
HTTPS_PROXY value with no scheme to be treated as http://.

Here we check the HTTPS_PROXY value, and if it has no scheme we reset
the value to prepend https://