QUICKSTART: prefer resolve:ssm: parameters for simplicity
The "simple option" was originally the note at the end of this section about
using the SSM web console.  That link changed, but we also discovered that you
have to have created a parameter in that region yourself to be able to use the
search quicklink.  Requiring that would be awkward, so we remove the link.

A simpler option for most people is to use the resolve:ssm: prefix to have EC2
dynamically fetch the latest AMI ID from SSM.  This updates the wording to
prefer that option, creating a new subsection for manually fetching parameters.