QUICKSTART: Remove BUILDING link
This was put in for completeness, but goes against the purpose of the
quickstart guides, which is to get an interested user launching Bottlerocket
quickly, and it's awkward to word such a different purpose here.

The BUILDING guide is linked from README right after the links to the
quickstart guides, for users that have more advanced needs and need to build
their own images.