host-ctr: support for bootstrap containers
This commit adds support for bootstrap containers in `host-ctr`. The
`container-type` flag was added to specify the type of container to be
setup, which defaults to `host`. This allows to add support for other
type of containers in the future without adding flags for each of them.

The root filesystem mounts were refactored out from `withSuperpowered`
to `withRootFilesystemMounts`, since bootstrap containers require access
to the underlying root filesystem without being given the superpowered
treatment.

Failed container tasks will now return an error if the task exited with
a non-zero status.