apiclient: rename update-related types with clearer prefix
Having types named "check", "apply", and "cancel" that are related to updates,
but don't have an "update" prefix, prevents us from having other types or
subcommands called "check", "apply", or "cancel".  This adds an "update" prefix
to each of them to make it clear what they're checking, applying, or canceling.