Add support for bootstrap-containers via settings
This commit adds support to create bootstrap containers through the
API. Bootstrap containers are host containers that can be used to setup
the host during the execution of the `configured` target.

These containers are created with the prefix `boot` to prevent
collisions with normal host containers. They can be setup to fail the
boot process if the underlying container task exists with a non-zero
status code.