ecs-settings-applier: remove bottlerocket.version attribute
This commit will remove the somewhat deceptive bottlerocket.version
attribute. The attributes described in ecs.config.json are set when the
instance first joins an ECS cluster, but are not updated in subsequent
runs of the ECS agent. This leads to updated Bottlerocket instances
continuing to show the lower version they were at first initialization
within the ECS API.