build: only build packages for current variant
Adds the packages that are needed to build a variant to the variant's
Cargo dependencies. Previously the build-variant makefile target assumed
that all packages were pre-built. Instead we now tell Cargo to build the
packages we need for the variant.

Creates a workspace in the variants directory and removes the workspace
in the packages directory. Updates the makefile accordingly. Now only
the packages that are part of the dependency tree starting with the
variant will be built.

The build-kmod-kit target previously depended on build-packages, which
is no longer available. So a new target, build-package, and a
specialization of it, build-kernel, have been added. These use the same
workspace context that is used when building variants to ensure that the
same rpm is used by build-kmod-kit.