aws-k8s: get cluster dns cidr from eks
We previously assumed that an EKS cluster DNS IP was either 10.100.0.10
or 172.20.0.10. This may be incorrect when the cluster has a custom
serviceIPv4CIDR setting.

Now we get the serviceIPv4CIDR from EKS describe-cluster so that we can
calculate the correct cluster DNS IP in the presence of serviceIPv4CIDR.