host-containers@: limit host container restarts
If a host container fails to start more than 5 times in a 5 minute interval,
the unit will no longer be restarted.

This way if the container is exiting quickly after each restart due to some
persistent error, we don't restart indefinitely.

Currently there is a 45 seconds delay between restarts. So if the container
exits within 15 seconds of it restarting 5 times in a row, we'll stop trying
to restart the service.

For other transient errors such as image pull errors (maybe the registry is
down for a few hours), the service should be allowed to restart until the
error is resolved. (host-ctr has expotential backoff implemented for
image pulls that maxes out after ~45 seconds of retries, so it'll never
exceed the restart burst limit)

Once the host-container unit enters a state where it no longer
automatically restarts, the user can still manually restart the unit
(through enabling and disabling the host-container via settings) and thus
restarting the restart cycle.