build: refactor os package dependencies
Signed-off-by: Ben Cressey <bcressey@amazon.com>

Change os.spec such that all first party code is installed when the os
package is installed. Change release.spec such that it requires only the
os package for first-party code instead of requiring each individual
first-party package.

This consolidates the package dependencies for first-party code into
one place, instead of picking some of the packages in "release" and
others through the variant definition.

This is part of #1361. We want to better manage package dependencies
with cargo so that we can build only the packages needed for a given
variant, and this change eliminates some one-off first-party package
dependencies that needed to be expressed at the variant level.