thar-be-settings: pass changed settings to restart commands
If a service is affected by changed settings, the changed settings are passed
to the restart commands of that service in an environment variable, separated
by spaces. The variable is only set if there are changed settings that affect
the service, and only the settings relevant to that service are included.

This commit adds a Services wrapper which is struct containing a map of service
name(String) to Service where Service has one field for changed_settings and another for
model::Service. This was added to remove the complications of different service
representations that were necessary for this change.