grub: do not strip files
This build produces a mix of host tools and target modules, and new
versions of RPM build scripts from Fedora will attempt to strip both.
When cross-compiling, there's no single version of `strip` that can
do this successfully, and this now fails when building on an x86_64
host for the aarch64 target.

We only use the tools and modules to generate the bootloader images,
and they are not shipped anywhere. Hence we do not really care about
extracting debuginfo and can just leave the files unstripped.

Signed-off-by: Ben Cressey <bcressey@amazon.com>