selinux-policy: update for kernel 5.10
Enable the new "genfs_seclabel_symlinks" capability to label symlinks
on kernel filesystems in the same way that files and directories are
labeled.

Add the new "perfmon", "bpf", and "checkpoint_restore" actions to the
"capability2" and "cap2_userns" classes. Add the new "perf_event" and
"lockdown" classes and the corresponding actions.

Add the new permissions into the blanket "systems" permission set, so
that we continue to rely on the existing kernel access checks without
adding new SELinux restrictions.

Signed-off-by: Ben Cressey <bcressey@amazon.com>