systemd: generate stable machine IDs for ARM KVM guests
Systemd doesn't recognize Amazon EC2 as a valid dmi vendor. The vendor
is used to detect the virtualization type, which then is used to
generate the machine ID. Since no dmi vendor is detected, systemd
defaults to use the random machine ID generator, which results in a
different machine ID after each boot. This patch extends the list of
valid dmi vendors, and adds `Amazon EC2` as a new vendor.