netdog: Persist current IP and add subcommand `node-ip` for retrieval
With this change, netdog will persist the node's current IP (read from
it's DHCP leases file) to /var/lib/netdog/current_ip.  It does this as
part of its "install" subcommand, which is run each time that wickedd
receives an update.  A new subcommand has also been added, which
retrieves the current IP from said file and returns it as JSON.  This
subcommand is intended for use as a settings generator.