Revert "apiserver: reap exited child processes"
This reverts commit 5c9eb7c0a5302405738aa47eba8e3a0c5f2116b4.

The waitpid thread was intercepting non-zombie children, because it waits on
pid -1 (any child), and does so very rapidly.  In some cases, this means things
like `Command::new().status()` that waitpid internally are racing the waitpid
thread, lose the race, and can't find their own child.

This was seen as apiserver (rarely) failing to launch thar-be-updates when an
updates API call came in; in fact, t-b-u would run, but apiserver couldn't
report its status and so had to fail.