Move pod-infra-container-image generation to schnauzer template helpers
Previously, pluto talked to IMDS to get the current region so it could generate
the right pause container URL.  This was added before we had the
settings.aws.region setting, so that was the only way.  However, it means extra
IMDS traffic, and it means you can't override the region of the image, which is
useful when testing new regions.  (It also means you can't override the image
arch, though that's less useful.)

This creates a 'pause-prefix' helper, like the existing ecr-prefix helper, that
can be used in schnauzer templates.

This also adds a 'goarch' helper, to convert Bottlerocket-standard architecture
names like x86_64 to "Go-like" architecture names like amd64, as used in the
pause container.