VMware: Force BIOS boot and hdd order
The vSphere CSI driver recommends the VMware Paravirtual SCSI
controller, and uses SCSI when attaching disks to VMs.  We observed the
BIOS prioritizes SCSI over NVME, so if a SCSI drive was attached using
this CSI driver and the machine was rebooted, it would attempt to boot
via the SCSI drive which obviously won't work.  This change forces the
BIOS to prioritize hdd nvme0:1 first.