selinux-policy: split rules for trusted subjects
Since 59264b48, privileged containers have received the `control_t`
label by default. This label is treated as a trusted subject, and has
broad powers of system configuration, including the ability to modify
files in /etc or change SELinux labels.

However, because these powers are granted by default, it creates the
risk of accidental misconfiguration, where files in /etc are modified
without realizing that these changes will be lost on reboot. Changing
SELinux labels can also happen unexpectedly, and could interfere with
the operation of the host or the reliability of upgrades.

To fix this, we split the rules for trusted subjects in two, to match
the expected access boundaries between privileged containers and the
underlying host.

Privileged subjects can still interact with all processes and read
all files, but can no longer change SELinux labels or edit files in
/etc. They also cannot unmount system mounts, or interact directly
with systemd or dbus.

Signed-off-by: Ben Cressey <bcressey@amazon.com>