early-boot-config: Change local_file into a function variants can use
Previously, `local_file` was an implementor of the
`PlatformDataProvider` trait.  This was a bit awkward, considering local
file is _not_ a platform.  It also meant it was a bit awkward to use
inside another provider.  This change makes `local_file` a function that
all platforms can use and makes use of it inside the `aws` module for
`aws-dev` variants.