host-containers: allow mount propagations from privileged containers
This commit adds support to propagate mount points created in
bootstrap and superpowered containers, across mount peer groups.

The root filesystem of bootstrap and superpowered containers is setup
with the `rshared` configuration to allow mounts propagations across
peer groups. All mount points attached to the containers are configured
as `rprivate` (except for the `mnt` mount). This prevents bootstrap and
superpowered containers from remounting directories in the host's root
filesystem.

The `/.bottlerocket/rootfs/mnt` mount point was added to bootstrap and
superpowered containers. This mount point is a bind mount that points to
`/mnt` in the host, which itself is a bind mount of `/local/mnt`. This
is required to let users create mount points underneath `/mnt`. This
mount point is setup with the `rshared` configuration to allow
propagations across peer groups. This is the only mount point from which
propagations are allowed across peer groups.

With this change, bootstrap containers now have access to all the
devices in the host. Also, they now have the `CAP_SYS_ADMIN` capability
to let users manage ephemeral disks. The logic to build the container
specs was refactored to provide a better understanding of what options
are set for the containers' spec.

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>