pubsys: Add VMware 'upload-ova' subcommand
This adds the subcommand `upload-ova` to `pubsys`.  This subcommand is
meant to be used via `cargo make` and allow a user to upload an OVA to
one or more VMware datacenters, optionally marking the VM as a template.
Under the hood, it accomplishes this by running `govc` in the SDK
container via Docker invocation, similar to how `buildsys` accomplishes
its task.

Like the rest of `pubsys`, it gets much of its configuration from
`Infra.toml`.  The configuration details all of the values needed to
communicate with a vSphere instance, and allows for users to specify
"common" configuration that multiple datacenters share.  Credentials for
different datacenters can be specified in a file located at
`~/.config/pubsys/vsphere-credentials.toml`.  Because we use `govc`
under the hood, at runtime we check for GOVC_* environment variables and
allow them to override VMware datacenter config in `Infra.toml` or
`vsphere-credentials.toml`

The subcommand allows a user to override the name of the resulting VM,
import spec, and specify a subset of datacenters than what's specified
in config.