Makefile: Add `upload-ova` and `vmware-template` targets
This change adds the ability to upload a built OVA to one or more VMware
software defined datacenters.  The `vmware-template` target uploads the
OVA and marks it as a VM template using the import spec.  Multiple
environment variables allow a user to override the VM name, list of
datacenters to upload to, and the import spec used to upload.