Add hostname configurability with setting: `settings.network.hostname`
This commit adds the ability to configure the hostname of the system via
a new setting `settings.network.hostname`.  If the setting isn't
populated, at boot time we use the settings generator `netdog
generate-hostname` to populate it.  The generator attempts to resolve
the current IP via reverse lookup, and if unsuccessful uses the IP to
create a hostname in the format "ip-x-x-x-x".

A new service `set-hostname` runs after `settings-applier` and writes
the hostname to `/proc/sys/kernel/hostname` via `netdog set-hostname
$HOSTNAME`.  The service unit file uses an env file that is a template
using the setting `settings.network.hostname` to populate the $HOSTNAME
variable.