Archive old migrations
See also: f5a48c582df43

These migrations have already been built and released with Bottlerocket
releases and don't need to be built again.  The build system and os package
currently build everything in the migrations directory, which can take a
significant amount of time with unneeded migrations.

Their source is still useful as a reference, and the files are small, so it's
worthwhile keeping them in-tree rather than having to search git history.