selinux-policy: drop external and unlabeled types
These types were mostly treated like `any_t` and `local_t`, where all
processes could freely write to the files. The special case was for
trusted processes, where directories created on an `unlabeled_t` path
would end up with the `local_t` label.

In combination with some mount options, this caused files on `/local`
to end up with the right labels, even if the filesystem was created
on a system without SELinux enabled. This might happen when using a
custom disk image as the source for the secondary storage volume.

However, a filesystem that's created by a bootstrap container won't
necessarily be mounted with the right options, and the `unlabeled_t`
label would continue to propagate. That would prevent the named file
transitions used to label Docker and containerd directories from
taking place, which would make them less secure.

We can simplify the policy and avoid this problem by treating unknown
or unrecognized types as already having the `local_t` label.

Signed-off-by: Ben Cressey <bcressey@amazon.com>