selinux-policy: add distinct type for container files
Previously we had two use cases for `local_t`. It was the label used
for most files and directories on `/local`, and therefore the label
that most hostPath mounts would have. It was also the label of the
container root filesystem, and therefore the label that external
volumes, emptyDir mounts, and other private storage would have.

For MCS isolation, it's useful to have distinct types to assert that
one type always has a level with categories, and another type never
does. That way, the constraints can be applied only to the files that
are meant to be private to a pod or container.

Signed-off-by: Ben Cressey <bcressey@amazon.com>