selinux-policy: use target's range for new files
If `defaultrange` is not specified in the policy, the lower part of
the range from the source process is applied to all new files.

Unprivileged containers will run with a process label that includes
two category pairs, so the files get the label we expect.

Privileged containers, on the other hand, may run with these labels:
* `system_u:system_r:control_t:s0-s0:c0.c1023`
* `system_u:system_r:super_t:s0`

In both cases, the lower range of the process is just `s0`, and files
would end up with that. This would allow unprivileged containers to
also modify the files.

We can avoid this by using the target's range instead, since Docker
and containerd CRI will ensure that volume mounts are labeled with
the appropriate range.

Signed-off-by: Ben Cressey <bcressey@amazon.com>