Add a hostname entry to `/etc/hosts` if hostname is unresolvable
Previously, `/etc/hosts` was a static file (in the `release` package)
that we copied into the image when building Bottlerocket.  In order to
have an entry in the file corresponding to hostname, a new "hosts"
service was added to the affected services for
`settings.network.hostname`.  This new "hosts" service has a templated
configuration file that uses `settings.network.hostname` to populate the
entry in `/etc/hosts` if the current hostname is unresolvable in DNS.
The templated configuration file uses the previously added
`add_unresolvable_hostname` template renderer helper.