glibc: update to 2.34
In 2.34, many of the separate libraries like `pthread` are integrated
in the main library, and shared objects are no longer installed with
versioned file names. Update the packaged file lists to reflect these
changes.

Add patches from the upstream branch for the 2.34 release, to pick up
additional fixes, such as the fix for CVE-2021-38604.

Sync the C.UTF-8 locale patch with recent changes from Fedora.

Signed-off-by: Ben Cressey <bcressey@amazon.com>