selinux-policy: implement MCS isolation
Docker and containerd's CRI plugin currently allocate two random
categories for all containers, and use them to label processes and
mounts. The intent is to add an additional layer of access control
between mutually untrusting containers. However, this is not enforced
by SELinux automatically; the policy needs rules to govern access.

This adds mlsconstrain rules to the policy so that these categories
actually have an effect. The constraints are based on the ones found
in the SELinux Reference Policy:
  https://github.com/SELinuxProject/refpolicy/blob/master/policy/mcs

They aren't exactly the same since our type attribute sets and class
permission sets are not as fine-grained. However, the broad strokes
are similar. File, directory, and process actions will be restricted
unless the subject's label dominates the target's label, which is
only true for targets in the same container, or for subjects that are
privileged.

All processes could previously set their existing label as the label
to use for newly created processes; this is now constrained so that
the level and categories cannot change unless the process is trusted.

Untrusted processes could not previously perform any relabel actions.
This is relaxed slightly to permit processes in containers to relabel
files in the container, but constrained so that the effective label
cannot change. This fixes programs (`pip`, `podman`) that attempt to
preserve the extended attributes of source files when copying them.

Signed-off-by: Ben Cressey <bcressey@amazon.com>