selinux-policy: relocate some label permissions
This moves certain "odd" permissions from the sets that control file
relabels and process transitions, to the more general set of actions
for managing system behavior.

The goal is to allow container processes to perform "no-op" relabels
and transitions, subject to policy constraints to ensure that the old
and new labels do not actually change.

The relabel actions for the `filesystem` class aren't constrained by
`mlsvalidatetrans` rules, so there's no way to assert that the label
would be unchanged afterwards.

It's not clear what the actions for the `kernel_service` class are
meant to do. The reference policy only allows them for `cachefilesd`,
so it should be fine to reserve them for trusted processes.

Signed-off-by: Ben Cressey <bcressey@amazon.com>