libglib: update to glib 2.70.0
Includes a fix for calling the `close_range()` function, which was
added in glibc 2.34.

glib2 still depends on the original PCRE library, rather than PCRE2
which we package as "libpcre". Older versions of glib2 included pcre
sources directly in the tree; newer versions use a meson subproject
and wrap to vendor the dependency.

meson is blocked from downloading sources, so we now include the pcre
archive and build overlay as external files. This package is expected
to remain the only consumer of PCRE in the distro.

Signed-off-by: Ben Cressey <bcressey@amazon.com>