logdog: Filter sensitive settings from API output
Previously, `logdog` used the `exec` mode to run `apiclient` and dump
everything from the datastore.  Since a few settings now exist that can
store sensitive data, we need to filter them from being written to disk.

In order to make this change, `logdog` needed to be trained to
understand the data it was gathering, in order to filter it.  A new
`settings` mode was added specifically for gathering settings.  This
mode uses the `apiclient` and `datastore` libraries to be able to
de/serialize the output, filter it, and then serialize it back to
disk.