wicked: add fix for dhcp6 defer timeout
We use a one second defer timeout for the DHCPv6 lease essentially to
mark it as optional and minimize the boot delay. One second is longer
than we would like already, but going sub-second is somewhat invasive
because the timeouts are tied to the protocol implementation and can
change the client behavior. It's relatively simple to avoid the extra
wait caused by an early timer event.

Signed-off-by: Ben Cressey <bcressey@amazon.com>