build: support custom OS and data image sizes
Add a new structure to variant metadata so the OS and data disk sizes
can be overridden. The previous values - 2 GiB for the OS disk, and 1
GIB for the data disk, are promoted to the default values so existing
variant definitions continue working.

References to "disk image" in `rpm2img` have been changed to refer to
"os image" instead, to avoid confusion over which image is meant. The
expected sizes for these images are now required parameters.

Signed-off-by: Ben Cressey <bcressey@amazon.com>