release: use systemd-repart for partition resizing
Switch from our `growpart` tool to `systemd-repart` to resize the
data partition. For a unified root+data image.

`growpart` uses the `gptman` crate, which calls the BLKRRPART ioctl
to tell the kernel to re-read the partition table. This call fails if
the device contains mounted partitions.

`systemd-repart` uses the newer BLKPG ioctl, which manipulates the
kernel's view of individual partitions. This works even if the root
filesystem is present on the same device and already mounted. It also
avoids the need to handle the partition symlink going away and coming
back, since udev does not get the change event that triggers this.

The two tools differ in how much free space is left on the device
after the last partition is resized. `growpart` ends the partition
one sector before the last 1 MiB boundary, while `systemd-repart`
ends it just before the GPT label. Both tools run on every boot.

To avoid problems on downgrade after a newer release resizes the data
filesystem beyond where the older release will end the partition, we
constrain `systemd-repart` to leave the older number of free sectors.

Since `/local` can be mounted during the resize operation, we can use
a real mount unit for it, which greatly simplifies the dependencies,
and allows us to decouple the "prepare" logic from "resize" logic.

Signed-off-by: Ben Cressey <bcressey@amazon.com>