systemd: exclude default repart service
The repart service is a oneshot, so we can't use a drop-in to replace
the `ExecStart` command that runs by default, which attempts to add
or grow defined partitions on the same device as the root filesystem.

This is OK for "unified" images, where the data partition is at the
end of the device, but not OK for "split" images, where it lives on a
different device.

Signed-off-by: Ben Cressey <bcressey@amazon.com>