selinux-policy: add rules to limit noatsecure access
Whenever applications that trigger transitions are called, SELinux will
set the `AT_SECURE` flag, which is used to clear various environment
variables that are considered harmful for a system (e.g. LD_PRELOAD).

By granting the `noatsecure` permission to processes, this environment
sanitation is disabled for a particular transition. Thus, we explicitly
forbid system subjects from using this permission, but we still allow
containers' processes to use it since we don't know if users require
the permission.

An additional rule was added to ignore the audit messages for programs
that triggered an implicit transition and don't have the `noatsecure`
permission.

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>