systemd: update to 250.4
Upstream reworked the VM detection logic for generating machine IDs,
and now detects the Amazon EC2 Nitro hypervisor. We still carry the
patch to generate machine IDs for Xen and VMware.

repart no longer depends on OpenSSL as of ade99252e, so we can drop
that patch.

Update the meson build options to make most choices explicit, and to
cover options added over the last few releases. This tidies up the
list of excluded files by not building components like oomd in the
first place.

Signed-off-by: Ben Cressey <bcressey@amazon.com>