release: fix prepare-local dependencies
In order for `setfiles` to work, the SELinux file contexts must have
been copied into `/etc`.

The dependency is specified with "Wants" rather than "Requires" to
avoid restarting the service if selinux-policy-files is restarted for
any reason. Subsequent runs would fail and put the system in a bad
state until the next reboot.

Add RefuseManualStop / RefuseManualStart to both services to indicate
the risk during interactive use by an administrator.

Signed-off-by: Ben Cressey <bcressey@amazon.com>