vmware-variants: migrate host-container versions to latest versions
We haven't been migrating host-container versions for vmware-variants.
This migration migrates all previous host-container versions to the
latest host-container version on upgrade.

It is unclear what version of the host-containers we should downgrade to
since it could be any of the older host-container versions.
We can just stay on the latest host-container version since there are no
breaking changes.