build: integrate bottlerocket-license-tool
Add support for the `Licenses.toml` file, used to provide licensing
information for some packages.
When the `Licenses.toml` file and the `licenses` folder are provided,
they are copied to the build directory of the package. If either path is
missing, empty configurations are provided. When this happens, packages
that use `Licenses.toml` will fail to build.

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>