early-boot-config: Make local file a valid source for all platforms
This change adds local file as a valid user data source for all
platforms we support.  Previously local file was only supported on
`aws-dev` and `metal-dev` variants.  A nice side affect of this change
is that it simplifies the platforms we support and streamlines the
conditional compilation a bit more.