glibc: update to 2.35
Add current patches from the release branch.

The C.UTF-8 locale is now upstream, so we no longer need the Fedora
patches that add it. Since it's not built-in, we need to generate and
install the corresponding files, which means building `localedef` for
the build host's architecture as well.

`--enable-static-pie` is now the default for architectures where it's
supported, so we no longer need to pass that option.

Exclude the new `ld.so` link since we don't need it, and `ldd` since
it's a shell script. `catchsegv` and `libSegFault.so` were removed
upstream.

Signed-off-by: Ben Cressey <bcressey@amazon.com>