release: Include a default .link file for network interfaces
This change adds a default network configuration .link file that
`systemd-udev` will use when configuring new interfaces.  It contains
the default list of policies that are used when naming interfaces, as
well as the policy by which the MAC address should be set.

Bottlerocket packages its own version of this file rather than the
default from systemd for a few reasons. 1) Bottlerocket does not
create/use a udev hwdb (we disable the option in systemd compile flags),
so we remove this option from the NamePolicy list, 2) CNI plugins can be
confused when MAC addresses change for virtual interfaces, so
Bottlerocket sets the default MACAddress Policy to "none" which
directs systemd not to attempt to manage the MAC.  Hardware usually has
a MAC, and veth devices used by CNI generally get a MAC generated by the
plugin.

Additional information about the MAC address issue:
https://github.com/systemd/systemd/issues/3374#issuecomment-288882355
https://github.com/flatcar-linux/Flatcar/issues/278
https://github.com/flatcar-linux/init/pull/33