netdog: Add ability to generate network configuration
This commit adds a subcommand `generate-net-config` to `netdog`.  It
uses a user-provided TOML file `net.toml` to generate the proper wicked
configuration files for the network interfaces in the system.  This file
lives on the BOTTLEROCKET-PRIVATE partition.

If no file is found, `netdog` looks at the kernel command line
(/proc/cmdline) for the `netdog.default-interface` prefix and configures
the interface listed there.  A single default interface may be defined
on the kernel command line with the format:
`netdog.default-interface=interface name:option1,option2`.
"interface name" is the name of the interface, and valid options are
"dhcp4" and "dhcp6".  A "?" may be added to the option to signify that
the lease for the protocol is optional and the system chouldn't wait for
it.  If the prefix does not exist on the kernel command line, no network
configuration is generated.

At configuration generation time, `netdog` writes the primary interface
to `/var/lib/netdog/primary_interface`.  Primary interface may be set in
`net.toml`; if it is absent in the file the first interface in the list
is assumed to be the primary.  If the interface is passed via kernel
command line, it is assumed to be the default.

Previously, the `netdog install` command, which is called by wicked for
configured interfaces, matched on the name of the interface (eth0)
before it took any action.  The `install` command writes the
resolv.conf, and the current_ip file.  For AWS/VMware we are explicitly
only configuring `eth0`, and for Metal variants we will not be able to
determine the configured interface names.  This commit changes the
behavior.  Rather than filtering on interface name in order to write
`resolv.conf` and the current IP file, we only write the files if the
interface currently being configured is the primary interface (read from
the previously written `primary_interface` file).

Note: Previously, our network configuration for `eth0` included a few
options, `arp-verify` and `arp-notify`.  These options were
(theoretically) meant to speed up boot time in EC2 since we know our IP
won't be a duplcate and therefore we don't need to ARP for it.  (That's
not how networking in EC2 works)  In our testing, `arp-very/notify` made
no significantly measureable difference in the time it took to get from
"lease request" to "committed lease".  From this commit on, these
options aren't part of our generated config for any variant.