netdog: Require `enabled` key in DHCP configuration
The change makes the `enabled` option required when setting DHCP
options.  Previously the `net.toml` would be parsed and wicked
configuration would be generated just fine, but the interface wouldn't
ever come up because DHCP 4/6 wasn't enabled.  This causes a string of
failures in later services that depend on network being up.  This change
ensures the config requires the `enabled` option be set.

Leaving the `enabled` option out entirely was considered, as one could
assume the user wants DHCP to be enabled if additional options were
being set.  However, we didn't want to assume this for users, as there
could possibly be a use case for setting options, but leaving DHCP
disabled for an interface, perhaps to be used later.