Move primary interface sysctl config to netdog
eth0 is not the default interface for all variants any longer.  This
change removes the hardcoded `eth0` specific sysctls from default
settings.  `netdog` is now responsible for setting these default sysctls
for whatever the primary interface is for the platform.

In order to do this, a new subcommand `prepare-primary-interface` is
added to `netdog` and run via a oneshot service.  The command reads the
primary interace from file (previously generated by `netdog`) and writes
the sysctls for that interface to file in `/etc/sysctl.d`.  It then
executes `systemd-sysctl`, which sets these sysctls.  Since all of this
happens early in boot, a user is able to override these via API settings
as usual.