k8s-1.23: enable `CSIMigration*` feature gates in kubelet
This enables the CSIMigration features gates that enables migration from
in-tree CSI plugins to solution-specific CSI drivers.

EKS 1.23 is planning to enable CSIMigrationAWS by default.

vSphere container storage plugin migration has been in beta since 1.19
and the in-tree plugin does not support the newest features in the
vSphere CSI driver.

Upstream kubernetes is eventually going to deprecate in-tree CSI
plugins.