host-ctr: propagate container storage mounts
Previously, the implementation anticipated that bootstrap containers
would use symlinks back to `/mnt` to redirect container storage to a
different mounted filesystem.

This turned out to be quite brittle in practice. Redirecting kubelet
volume storage in that way was found to be fundamentally incompatible
with CSI driver expections, which require the "real" mount point to
be `/var/lib/kubelet` in order to work correctly.

To accommodate this, container storage directories are now treated as
a special case. If they exist in the host, they are mounted into the
bootstrap container with the "rbind,rshared" options so that changes
will propagate back to the host.

Signed-off-by: Ben Cressey <bcressey@amazon.com>