release: split up /local preparation
Rather than setting up `/mnt`, `/opt`, and `/var` in the same unit,
switch to dedicated units that prepare the corresponding mount. This
lets the system do more in parallel, and makes it trivial to reason
about when `/local` is fully ready and done - which it now is as soon
as it's mounted.

Signed-off-by: Ben Cressey <bcressey@amazon.com>