host-containers: don't forcibly clean-up right after disabling service
After disabling the host-container's systemd service unit, we need to
allow it ample time for it to actually gracefully exit. We shouldn't
just immediately try to kill the container process after stopping the
host-container service. `host-ctr` should already properly handle the
termination signal propagation to the container task when stopping the
systemd unit normally. `host-ctr` also has a kill switch for when the
container task does not exit in time. Therefore, it's highly unlikely
that a host-container process will linger long after we disable said
host-container.