os: rework host and bootstrap container reloads
Since there's an implied systemd reload by default when enabling or
disabling a service, a reload was triggered for each bootstrap and
host container.

Every systemd reload costs at least 200 ms, which is a lot to pay
during boot when the services can't even be running yet.

For host containers, expand the conditional logic around whether the
system is fully booted to cover the "enabled" and "disabled" cases,
and skip the reload unless the boot is complete. Simplify the logic
by unconditionally stopping and cleaning up each container first, to
clarify the actual difference in behavior.

For bootstrap containers, skip the reload in all cases, since systemd
will be reloaded when switching the default to `configured.target`.

Signed-off-by: Ben Cressey <bcressey@amazon.com>