variants: Enable the metal-dev build for the Arm architecture
For testing purposes, the metal-dev variant can be built and runs just
fine on the Arm architecture (tested with a QEMU/KVM virtual machine on
Amazon EC2 Graviton instances). Remove the restriction to x86_64 from
the variant definition.

Note that this keeps the restriction for the non-development metal
variants for now.

Signed-off-by: Markus Boehme <markubo@amazon.com>