kubelet: load ipvs modules for kube-proxy
The `insmod` binary that kube-proxy uses does not support compressed
modules, and hands the compressed file to the kernel to load. This
leads to warnings that unsigned modules are not allowed, if lockdown
is enabled, or errors about malformed modules otherwise.

kube-proxy does check whether the module is loaded, so we can avoid
these errors by loading the necessary modules before kubelet starts.

Signed-off-by: Ben Cressey <bcressey@amazon.com>