buildsys: support scoped variant sensitivity
The previous definition of variant sensitivity treated any change to
the variant as requiring a rebuild.

However, packages might depend on only one component of the variant
tuple. For example, the kernel package might conditionally enable
some modules for the "aws" platform, and a different set for the
"metal" platform.

To prevent unnecessary rebuilds, packages are now allowed to specify
which component, or group of components, they depend upon.

Signed-off-by: Ben Cressey <bcressey@amazon.com>