netdog: Add new `write-resolv-conf` subcommand
This commit adds a new `write-resolv-conf` subcommand, which is meant to
be used as a restart command for DNS API settings.  It will write the
`resolv.conf` using DNS settings if they exist, supplementing any
missing settings with settings from the DHCP lease of the primary
interface if it exists.