tools: add diff-kernel-config to identify kernel config changes
With multiple kernel versions being supported for two architectures in
cloud and metal flavors it can be tedious and error-prone to get the
full picture on the effects changes have on the resulting kernel
configurations. The new diff-kernel-config tool helps automate this
task.

diff-kernel-config compares kernel configurations before and after a
series of commits to the Bottlerocket repository. For this, it runs
several full kernel builds that cover all possible combinations of
kernel version, architecture, and cloud/metal flavor that could be
shipped in any of the official variants. Combinations not used in any
variant are skipped.

diff-kernel-config produces kernel config diffs broken down by
combination in its output directory, alongside a comprehensive report of
all changes and a summary that can aid the code review process.

Signed-off-by: Markus Boehme <markubo@amazon.com>