macros: add -Bsymbolic-functions to ldflags
In 237946dc, `-fno-semantic-interposition` was added to the default
build flags. This allows the compiler to assume that exported symbols
will not be interposed and to optimize accordingly, for example by
inlining functions.

`-Bsymbolic-functions` tells the linker to bind references to global
function symbols to the definition inside the shared library, if any,
which should be safe given that the compiler has already optimized
the code under the assumption that those definitions will be used.

Signed-off-by: Ben Cressey <bcressey@amazon.com>