netdog: Add constructor for `WickedInterface`
This commit adds a `new()` constructor for `WickedInterface`.  This
constructor helps to hide defaults for `WickedControl`, and provides a
base on which to build out the structure (in lieu of a full on builder).
As we add additional struct members, we can add them to this method
rather than to all the callers/users of `WickedInterface`.