netdog: Add macros to simplify and improve net config testing
This commit adds a new module `test_macros` inside of `net_config`.  The
module is meant to contain a group of declarative macros, each of which
contain a scoped group of tests and can be used across net config
versions.  The macros themselves are declarative and define a new module
that contains all the tests.  This allows us to use the same tests
across all versions of net config, and get nicely formatted `cargo test`
output showing the net config version, module, and test.  This is much
nicer than looping over versions within a single test, and provides much
better error output for the user.